Introduce `tslint` to lint for TypeScript (#206)
Currently, we are using `clang-format --diff` command to lint for
TypeScript. It's useful but not tightly linter. So, we should introduce
`tslint` to lint for TypeScript tightly.

This change applies the linter to gulpfile.ts once. And then we will
apply the linter to other files in the future.

ISSUE=#205